fix guest path, disable path checking on other backing stores
Serge Hallyn recommends using relative paths for mountpoints in
containers. It prevents problems with cloning, as well as when using
backingstores that aren't visible in the host file tree. The old path
wouldn't work on lvm, for instance.

There's no check for the directory on other backingstores yet. I am
using this with LVM, and I don't see a way to check or create it in this
configuration.